Building a DAO layer

+++ Chapter04/MySQL/Maven/JpaSimpleGenericDao/src/main/java/com/classicmodels/jooq/repository/SaleRepository.java
@@ -10,5 +10,5 @@ import org.springframework.transaction.annotation.Transactional;
 public interface SaleRepository extends ClassicModelsRepository<Sale, Long> {
 
     public List<Sale> findSaleByFiscalYear(int year);
-    public List<Sale> findSaleAscGtLimit(int limit);
+    public List<Sale> findSaleAscGtLimit(double limit);
 }
